feat: router, model 인스턴스 생성을 app.js로 위임하여 controller의 복잡도 저감
