+++ src/controllers/main.js
@@ -1,7 +1,7 @@
 import renderMainPage from '../views/main/index';
 
-const setClickEventListenerOnArticle = router => {
-  const $articleHeadContainers = document.querySelectorAll('li');
+const onClickArticle = router => {
+  const $articleHeadContainers = document.querySelectorAll('li'); // 추후 CSS 작업 이후 class가 적용된 다음 다시 변경 예정 Ticket: WOO-83
   $articleHeadContainers.forEach($li => {
     $li.addEventListener('click', () => {
       window.history.pushState('', '', `/article/${$li.id}`);
@@ -16,7 +16,7 @@ function Main(router, model) {
     try {
       const contents = await model.getArticleList();
       renderMainPage(contents);
-      setClickEventListenerOnArticle(router);
+      onClickArticle(router);
     } catch (e) {
       if (import.meta.env.DEV) console.error(`Error at controllers/main.js render : ${e}`);
     }
